Move `bench`-folder in `tests`-folder
First part of reorganizing and cleaning up test-folders.
