+++ tests/bench/templates/arguments.js
