+++ tests/bench/templates/array-each.js
