+++ tests/bench/templates/array-mustache.js
