+++ tests/bench/templates/complex.handlebars
