+++ tests/bench/templates/complex.js
