+++ tests/bench/templates/data.js
