+++ tests/bench/templates/depth-1.js
