+++ tests/bench/templates/depth-2.js
