+++ tests/bench/templates/index.js
