+++ tests/bench/templates/object-mustache.js
