+++ tests/bench/templates/object.js
