+++ tests/bench/templates/partial-recursion.js
