+++ tests/bench/templates/partial.js
