+++ tests/bench/templates/paths.js
