+++ tests/bench/templates/string.js
