+++ tests/bench/templates/variables.js
