+++ tests/bench/util/template-runner.js
